fix: type fixes & folder structure
