+++ src/features/AddToCart/model/types.ts
@@ -1,0 +1,3 @@
+export interface IAddBtnProps {
+  size: string;
+}
