--- src/widgets/CartItemList/model/CartItemList.interface.ts
@@ -1,8 +1,0 @@
-export interface ICartItem {
-  id: string;
-  picture: string;
-  name: string;
-  price: string;
-  inStock: string;
-  size: string;
-}
